Move tests to features layer
